+++ src/app/components/sidebar-tokens.js
@@ -2,13 +2,11 @@
  * A sidebar for tokens (extension of `two-level-sidebar`)
  *
  * @module components/sidebar-tokens
- * @author Michal Borzecki
- * @copyright (C) 2018 ACK CYFRONET AGH
+ * @author Michał Borzęcki
+ * @copyright (C) 2019 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
-import { computed } from '@ember/object';
-import { A } from '@ember/array';
 import TwoLevelSidebar from 'onedata-gui-common/components/two-level-sidebar';
 import layout from 'onedata-gui-common/templates/components/two-level-sidebar';
 
@@ -24,29 +22,11 @@ export default TwoLevelSidebar.extend({
 
   /**
    * @override
-   * Keep original sorting from server, but make a copy of array
    */
-  sortedCollection: computed('model.collection.list.[]', function () {
-    return A(this.get('model.collection.list').toArray());
-  }),
-
-  /**
-   * @override
-   */
-  firstLevelItemIcon: 'tokens',
+  firstLevelItemComponent: 'sidebar-tokens/token-item',
 
   /**
    * @override
    */
   sidebarType: 'tokens',
-
-  /**
-   * @override
-   */
-  secondLevelItems: Object.freeze([]),
-
-  /**
-   * @override
-   */
-  showCreateOnEmpty: false,
 });