+++ src/app/locales/en/components/remove-token-modal.js
@@ -1,0 +1,6 @@
+export default {
+  headerText: 'Remove token',
+  messageText: 'Are you sure you want to remove token "{{tokenName}}"?',
+  proceed: 'Remove',
+  cancel: 'Cancel',
+};