+++ src/app/locales/en/services/token-actions.js
@@ -2,4 +2,6 @@ export default {
   createToken: 'Create new token',
   tokenCreateSuccess: 'Token created successfully.',
   tokenCreation: 'token creation',
+  tokenRemoveSuccess: 'Token has been removed successfully.',
+  removingToken: 'removing token',
 };