+++ src/app/models/group.js
@@ -14,7 +14,7 @@ import { alias, equal } from '@ember/object/computed';
 
 import StaticGraphModelMixin from 'onedata-gui-websocket-client/mixins/models/static-graph-model';
 import GraphSingleModelMixin from 'onedata-gui-websocket-client/mixins/models/graph-single-model';
-import InvitingModelMixin from 'onedata-gui-websocket-client/mixins/models/inviting-model';
+import InvitingModelMixin from 'onezone-gui/mixins/models/inviting-model';
 
 export default Model.extend(GraphSingleModelMixin, InvitingModelMixin, {
   onedataGraphUtils: service(),