+++ src/app/services/sidebar-resources.js
@@ -2,29 +2,29 @@
  * An abstraction layer for getting data for sidebar of various tabs
  *
  * @module services/sidebar-resources
- * @author Jakub Liput, Michal Borzecki
- * @copyright (C) 2017-2018 ACK CYFRONET AGH
+ * @author Jakub Liput, Michał Borzęcki
+ * @copyright (C) 2017-2019 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
-import { inject } from '@ember/service';
+import { inject as service } from '@ember/service';
 import { A } from '@ember/array';
 import { resolve, reject } from 'rsvp';
 import SidebarResources from 'onedata-gui-common/services/sidebar-resources';
 
 export default SidebarResources.extend({
-  providerManager: inject(),
-  clientTokenManager: inject(),
-  clientTokenActions: inject(),
-  currentUser: inject(),
-  spaceManager: inject(),
-  clusterActions: inject(),
-  spaceActions: inject(),
-  groupManager: inject(),
-  groupActions: inject(),
-  clusterManager: inject(),
-  harvesterManager: inject(),
-  harvesterActions: inject(),
+  providerManager: service(),
+  tokenManager: service(),
+  tokenActions: service(),
+  currentUser: service(),
+  spaceManager: service(),
+  clusterActions: service(),
+  spaceActions: service(),
+  groupManager: service(),
+  groupActions: service(),
+  clusterManager: service(),
+  harvesterManager: service(),
+  harvesterActions: service(),
 
   /**
    * @param {string} type
@@ -37,7 +37,7 @@ export default SidebarResources.extend({
       case 'clusters':
         return this.get('clusterManager').getClusters();
       case 'tokens':
-        return this.get('clientTokenManager').getClientTokens();
+        return this.get('tokenManager').getTokens();
       case 'spaces':
         return this.get('spaceManager').getSpaces();
       case 'groups':
@@ -63,7 +63,7 @@ export default SidebarResources.extend({
       case 'clusters':
         return this.get('clusterActions.buttons');
       case 'tokens':
-        return this.get('clientTokenActions.actionButtons');
+        return this.get('tokenActions.actionButtons');
       case 'spaces':
         return this.get('spaceActions.buttons');
       case 'groups':
@@ -72,6 +72,17 @@ export default SidebarResources.extend({
         return this.get('harvesterActions.buttons');
       default:
         return [];
+    }
+  },
+
+  /**
+   * @override
+   */
+  getItemsSortingFor(type) {
+    if (type === 'tokens') {
+      return ['isActive:desc', 'name'];
+    } else {
+      return this._super(...arguments);
     }
   },
 });